+++ src/constants/flattenedIdentifiers.js
@@ -1,0 +1,8 @@
+const flattenedIdentifiers = [
+  { type: 'ISSN', subtype: 'Online' },
+  { type: 'ISSN', subtype: 'Print' },
+  { type: 'ISBN', subtype: 'Online' },
+  { type: 'ISBN', subtype: 'Print' },
+];
+
+export default flattenedIdentifiers;