+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add tests for SettingsRoute component. (UIEH-1198)
 * Add tests for SettingsKnowledgeBaseRoute component. (UIEH-1196)
+* Add tests for TitleCreateRoute component. (UIEH-1199)
 
 
 ## [7.0.1] (https://github.com/folio-org/ui-eholdings/tree/v7.0.1) (2021-10-19)